Rearrange folders and use story files instead
