+++ src/components/ExampleBox.stories.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import { action } from '@storybook/addon-actions';
+
+import ExampleBox from './ExampleBox';
+
+export default {
+  title: 'Design System|Example Box',
+  component: ExampleBox,
+};
+
+export const test = () => <ExampleBox onClick={action('clicked')} />;
+
+test.story = {
+  name: 'this is a test component',
+};
